move pusher event, subscribe when add email clicked and unsubscribe when event handled
