+++ src/libs/actions/Report.js
@@ -650,15 +650,6 @@ function updateReportWithNewAction(
 }
 
 /**
- * Updates isFromPublicDomain in Onyx.
- *
- * @param {Boolean} isFromPublicDomain
- */
-function setIsFromPublicDomain(isFromPublicDomain) {
-    Onyx.merge(ONYXKEYS.USER, {isFromPublicDomain});
-}
-
-/**
  * Updates a report in Onyx with a new pinned state.
  *
  * @param {Number} reportID
@@ -747,26 +738,6 @@ function subscribeToUserEvents() {
                 '[Report] Failed to subscribe to Pusher channel',
                 false,
                 {error, pusherChannelName, eventName: Pusher.TYPE.REPORT_TOGGLE_PINNED},
-            );
-        });
-
-    // Live-update if a user has private domains listed as primary or secondary logins.
-    Pusher.subscribe(pusherChannelName, Pusher.TYPE.ACCOUNT_VALIDATED, (pushJSON) => {
-        Log.info(
-            `[Report] Handled ${Pusher.TYPE.ACCOUNT_VALIDATED} event sent by Pusher`,
-            false,
-            {isFromPublicDomain: pushJSON.isFromPublicDomain},
-        );
-        setIsFromPublicDomain(pushJSON.isFromPublicDomain);
-    }, false,
-    () => {
-        NetworkConnection.triggerReconnectionCallbacks('pusher re-subscribed to private user channel');
-    })
-        .catch((error) => {
-            Log.info(
-                '[Report] Failed to subscribe to Pusher channel',
-                false,
-                {error, pusherChannelName, eventName: Pusher.TYPE.ACCOUNT_VALIDATED},
             );
         });
 }
